fix floating issues

+++ app/routes/refunds.py
@@ -1,4 +1,6 @@
 import json
+from decimal import Decimal
+
 from flask import Blueprint, flash, redirect, render_template, request, session, url_for
 
 from ..utils.api import make_api_request, process_intent
@@ -41,7 +43,7 @@ def unlinked_refund():
             flash(error or "Invalid amount format", "danger")
             return redirect(url_for("refunds.unlinked_refund"))
 
-        amount = float(amount_str)
+        amount = Decimal(amount_str)
         merchant_reference = request.form.get("merchant_reference")
         if not merchant_reference:
             flash("Merchant reference is required", "danger")
@@ -98,7 +100,7 @@ def linked_refund():
             flash(error or "Invalid amount format", "danger")
             return redirect(url_for("refunds.linked_refund"))
 
-        amount = float(amount_str)
+        amount = Decimal(amount_str)
         merchant_reference = request.form.get("merchant_reference")
         if not merchant_reference:
             flash("Merchant reference is required", "danger")
